pyproject.toml
[tool.poetry]
name = "python-mobile-test-template"
version = "0.1.0"
description = "fully charged mobile end-to-end testing framework with selene + appium + pytest"
authors = ["yashaka <yashaka@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "python_mobile_test_template"}]

[tool.poetry.dependencies]
python = "^3.10"
selene = {version = "^2.0.0-beta.8", allow-prereleases = true}
Appium-Python-Client = "^2.6.1"
pytest = "^7.1.3"
pytest-xdist = "^2.5.0"
python-dotenv = "^0.21.0"
pydantic = "^1.10.2"

[tool.poetry.dev-dependencies]
black = "^22.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 79
target-version = ['py310']
skip-string-normalization = true
